Move functions from FocallyLiftedEndcap into new file.
No functions are changed, only moved.
These functions will be needed for other maps, so this
is why they were factored out.
