+++ src/Domain/CoordinateMaps/CMakeLists.txt
@@ -11,6 +11,7 @@ spectre_target_sources(
   Affine.cpp
   BulgedCube.cpp
   CylindricalEndcap.cpp
+  CylindricalEndcapHelpers.cpp
   CylindricalFlatEndcap.cpp
   CylindricalFlatSide.cpp
   CylindricalSide.cpp
@@ -43,6 +44,7 @@ spectre_target_headers(
   CoordinateMap.tpp
   CoordinateMapHelpers.hpp
   CylindricalEndcap.hpp
+  CylindricalEndcapHelpers.hpp
   CylindricalFlatEndcap.hpp
   CylindricalFlatSide.hpp
   CylindricalSide.hpp
