+++ src/Domain/CoordinateMaps/FocallyLiftedEndcap.cpp
@@ -9,6 +9,7 @@
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
+#include "Domain/CoordinateMaps/CylindricalEndcapHelpers.hpp"
 #include "Domain/CoordinateMaps/FocallyLiftedMapHelpers.hpp"
 #include "Parallel/PupStlCpp11.hpp"
 #include "Utilities/ConstantExpressions.hpp"
@@ -19,78 +20,6 @@
 #include "Utilities/MakeWithValue.hpp"
 
 namespace domain::CoordinateMaps::FocallyLiftedInnerMaps {
-
-namespace Endcap_detail {
-double sin_ax_over_x(const double x, const double a) {
-  // sin(ax)/x returns the right thing except if x is zero, so
-  // we need to treat only that case as special.
-  return x == 0.0 ? a : sin(a * x) / x;
-}
-DataVector sin_ax_over_x(const DataVector& x, const double a) {
-  DataVector result(x);
-  for (size_t i = 0; i < x.size(); ++i) {
-    result[i] = sin_ax_over_x(x[i], a);
-  }
-  return result;
-}
-double one_over_x_d_sin_ax_over_x(const double x, const double a) {
-  // Evaluates (1/x) d/dx [ sin(ax)/x ], which is the quantity that
-  // approaches a finite limit as x approaches zero.
-  //
-  // Here we need to worry about roundoff. Note that we can expand
-  // (1/x) d/dx [ sin(ax)/x ] =
-  // a/x^2 [ 1 - 1 - 2(ax)^2/3! + 4(ax)^4/5! - 6(ax)^6/7! + ...],
-  // where I kept the "1 - 1" above as a reminder that when evaluating
-  // this function directly as (a * cos(ax) - sin(ax) / x) / square(x)
-  // there can be significant roundoff because of the "1" in each of the
-  // two terms that are subtracted.
-  //
-  // The relative error in the above expression, if evaluated directly,
-  // is 3*eps/(ax)^2 where eps is machine epsilon.  (That expression comes
-  // from replacing "1 - 1" with eps and noting that the correct answer
-  // is the 2(ax)^2/3! term and eps is an error contribution).
-  // This means the error is 100% if (ax)^2 is 3*eps.
-  //
-  // The solution is to evaluate the series if (ax) is small enough.
-  // Suppose we keep up to and including the (ax)^(2n) term in the
-  // series.  Then the series is accurate if the (ax)^{2n+2} term (the
-  // next term in the series) is small, i.e. if
-  // (2n+2)(ax)^{2n+2}/(2n+3)! < eps.
-  //
-  // For the worst case of (2n+2)(ax)^{2n+2}/(2n+3)! == eps, the direct
-  // evaluation still has a relative error of 3*eps/(ax)^2, which evaluates to
-  // error = 3*eps* eps^{-1/(n+1)} * ((2n+2)/(2n+3)!)^{1/(n+1)}.
-  // This can be rewritten as
-  // error = 3 * [eps^n*((2n+2)/(2n+3)!)]^{1/(n+1)}.
-  //
-  // For certain values of n:
-  // n=1    error=3*sqrt(eps/30)               ~ 5e-9
-  // n=2    error=3*(eps^2/840)^(1/3)          ~ 7e-12
-  // n=3    error=3*(eps^3/45360)^(1/4)        ~ 2e-13
-  // n=4    error=3*(eps^4/3991680)^(1/5)      ~ 2e-14
-  // n=5    error=3*(eps^5/518918400)^(1/6)    ~ 5e-15
-  // n=6    error=3*(eps^6/93405312000)^(1/7)  ~ 1e-15
-  //
-  // We gain less and less with each order.
-  //
-  // So here we choose n=3.
-  // Then the series above can be rewritten
-  // 1/x d/dx [ sin(ax)/x ] = a/x^2 [- 2(ax)^2/3! + 4(ax)^4/5! - 6(ax)^6/7!]
-  //                        = -a^3/3 [ 1 - 4*3*(ax)^2/5! + 6*3*(ax)^4/7!]
-  const double ax = a * x;
-  return pow<8>(ax) < 45360.0 * std::numeric_limits<double>::epsilon()
-             ? (-cube(a) / 3.0) *
-                   (1.0 + square(ax) * (-0.1 + square(ax) / 280.0))
-             : (a * cos(ax) - sin(ax) / x) / square(x);
-}
-DataVector one_over_x_d_sin_ax_over_x(const DataVector& x, const double a) {
-  DataVector result(x);
-  for (size_t i = 0; i < x.size(); ++i) {
-    result[i] = one_over_x_d_sin_ax_over_x(x[i], a);
-  }
-  return result;
-}
-}  // namespace Endcap_detail
 
 Endcap::Endcap(const std::array<double, 3>& center, const double radius,
                const double z_plane)
@@ -124,7 +53,7 @@ void Endcap::forward_map(
   // Use z and y as temporary storage to avoid allocations,
   // before setting them to their actual values.
   z = sqrt(square(xbar) + square(ybar));
-  y = Endcap_detail::sin_ax_over_x(z, theta_max_) * radius_;
+  y = cylindrical_endcap_helpers::sin_ax_over_x(z, theta_max_) * radius_;
   x = y * xbar + center_[0];
   y = y * ybar + center_[1];
   z = radius_ * cos(z * theta_max_) + center_[2];
@@ -151,11 +80,11 @@ void Endcap::jacobian(const gsl::not_null<tnsr::Ij<tt::remove_cvref_wrap_t<T>,
   // Use parts of Jacobian as temp storage to reduce allocations.
   get<1, 1>(*jacobian_out) = sqrt(square(xbar) + square(ybar));
   get<1, 0>(*jacobian_out) =
-      radius_ *
-      Endcap_detail::sin_ax_over_x(get<1, 1>(*jacobian_out), theta_max_);
+      radius_ * cylindrical_endcap_helpers::sin_ax_over_x(
+                    get<1, 1>(*jacobian_out), theta_max_);
   // 1/rhobar d/d(rhobar) [(sin (rhobar theta_max)/rhobar)]
   get<0, 1>(*jacobian_out) =
-      radius_ * Endcap_detail::one_over_x_d_sin_ax_over_x(
+      radius_ * cylindrical_endcap_helpers::one_over_x_d_sin_ax_over_x(
                     get<1, 1>(*jacobian_out), theta_max_);
 
   // dy/dybar
@@ -205,12 +134,13 @@ void Endcap::inv_jacobian(
   get<1, 0>(*inv_jacobian_out) = sqrt(square(xbar) + square(ybar));
 
   // q = sin(rhobar theta)/rhobar as defined in the documentation, Eq. 17.
-  get<0, 1>(*inv_jacobian_out) =
-      Endcap_detail::sin_ax_over_x(get<1, 0>(*inv_jacobian_out), theta_max_);
+  get<0, 1>(*inv_jacobian_out) = cylindrical_endcap_helpers::sin_ax_over_x(
+      get<1, 0>(*inv_jacobian_out), theta_max_);
 
   // 1/rhobar dq/d(rhobar)
-  get<1, 1>(*inv_jacobian_out) = Endcap_detail::one_over_x_d_sin_ax_over_x(
-      get<1, 0>(*inv_jacobian_out), theta_max_);
+  get<1, 1>(*inv_jacobian_out) =
+      cylindrical_endcap_helpers::one_over_x_d_sin_ax_over_x(
+          get<1, 0>(*inv_jacobian_out), theta_max_);
 
   // Right-hand side of Eq. 23, without the factor of
   // xbar ybar/rq or the minus sign.
